* ext/sdbm/_sdbm.c (sdbm_open): use size_t.
* ext/syck/bytecode.c: ditto.

* ext/sdbm/_sdbm.c (delpair): use ptrdiff_t.

* ext/sdbm/init.c: use RSTRING_LENINT.

* ext/dl/handle.c: suppress warning: shorten-64-to-32.

* ext/strscan/strscan.c: ditto.

* ext/syck/emitter.c: ditto.

* ext/syck/implicit.c: ditto.

* ext/syck/syck.c: ditto.

* ext/syck/token.c: ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@31176 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/syck/bytecode.c
@@ -325,7 +325,7 @@ yy24:
             CHK_NL(YYCURSOR);
             if ( qstr[0] == '!' )
             {
-                int qidx = strlen( qstr );
+                size_t qidx = strlen( qstr );
                 if ( qstr[1] == '\0' )
                 {
                     free( qstr );
